requirements.txt
python-telegram-bot>=20.7,<21.0
httpx>=0.26.0
Pillow>=10.2.0
